KCALCULATOR/requirements.txt
asgiref==3.3.4
beautifulsoup4==4.9.3
cycler==0.10.0
Django==3.2
django-bootstrap-v5==1.0.2
django-bootstrap4==3.0.0
djangorestframework==3.12.4
kiwisolver==1.3.1
matplotlib==3.4.1
numpy==1.20.2
pandas==1.2.4
Pillow==8.2.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
response==0.5.0
scipy==1.6.2
six==1.15.0
soupsieve==2.2.1
sqlparse==0.4.1
